refactor: move groupmembershipclaim to application pkg
