+++ pkg/azure/util/application_builder.go
@@ -5,7 +5,7 @@ import (
 	msgraph "github.com/nais/msgraph.go/v1.0"
 
 	"github.com/nais/azureator/pkg/azure"
-	"github.com/nais/azureator/pkg/azure/util/groupmembershipclaim"
+	"github.com/nais/azureator/pkg/azure/client/application/groupmembershipclaim"
 )
 
 type ApplicationBuilder struct {
